requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
requests==2.31.0 